# academy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Summer Academy educational platform",
        "title": "Summer Academy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/terminal/{id}": {
            "get": {
                "security": [
                    {
                        "JWTCookie": []
                    }
                ],
                "description": "Renders the terminal interface for an active terminal session",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "terminal"
                ],
                "summary": "Display terminal interface",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Terminal session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Terminal page"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/terminal/{slug}": {
            "post": {
                "security": [
                    {
                        "JWTCookie": []
                    }
                ],
                "description": "Creates a new WBFY terminal session for a problem",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "terminal"
                ],
                "summary": "Create a new terminal session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Problem slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Programming language (default: bash)",
                        "name": "language",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Terminal session created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service unavailable - no ports available",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "description": "Bearer JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "JWTCookie": {
            "description": "JWT token stored in cookie for authentication",
            "type": "apiKey",
            "name": "academy_session",
            "in": "cookie"
        }
    }
}